[enhancement] Added a demoscene with a timer and resolved issues with a scene timer

+++ src/main/java/nl/meron/yaeger/engine/Timer.java
@@ -8,7 +8,7 @@ package nl.meron.yaeger.engine;
  */
 public abstract class Timer {
 
-    private int interval;
+    private long interval;
     private long prevTime = 0;
 
     /**
